+++ binding.gyp
@@ -7,7 +7,8 @@
                 "src/ons_options.cpp",
                 "src/consumer_ack.cpp",
                 "src/consumer.cpp",
-                "src/producer.cpp"
+                "src/producer.cpp",
+                "src/consumer_listener.cpp"
             ],
 
             "include_dirs": [
