+++ lib/consumer.js
@@ -15,7 +15,6 @@ if(os.platform() === "darwin") {
 }
 var _Consumer = require("../build/Release/ons").ONSConsumer;
 
-
 /**
  * Consumer
  * @param {String} cid the consumer id
@@ -78,12 +77,12 @@ Consumer.prototype.listen = function() {
 /**
  * stop
  */
-Consumer.prototype.stop = function() {
-    this.consumer.stop();
-
+Consumer.prototype.stop = function(callback) {
     if(this.keepTimer) {
         clearInterval(this.keepTimer);
     }
+
+    this.consumer.stop(callback);
 };
 
 module.exports = Consumer;
