+++ lib/producer.js
@@ -51,12 +51,12 @@ Producer.prototype.start = function(callback) {
 /**
  * stop
  */
-Producer.prototype.stop = function() {
-    this.producer.stop();
-
+Producer.prototype.stop = function(callback) {
     if(this.keepTimer) {
         clearInterval(this.keepTimer);
     }
+
+    this.producer.stop(callback);
 };
 
 /**
