+++ src/producer.cpp
@@ -18,135 +18,14 @@
 #include "producer.h"
 #include "ONSClientException.h"
 
+#include "producer_workers/producer_prepare_worker.h"
+#include "producer_workers/producer_send_worker.h"
+#include "producer_workers/producer_stop_worker.h"
+
 std::string producer_env_v = std::getenv("NODE_ONS_LOG") == NULL ?
         "" : std::getenv("NODE_ONS_LOG");
 
 Nan::Persistent<v8::Function> ONSProducerV8::constructor;
-
-class ProducerPrepareWorker : public Nan::AsyncWorker {
-public:
-    ProducerPrepareWorker(Nan::Callback* callback, ONSProducerV8& ons) :
-        AsyncWorker(callback),
-        ons(ons),
-        factory_info(ons.factory_info)
-    {
-    }
-
-    ~ProducerPrepareWorker() {}
-
-    void Execute()
-    {
-        real_producer = ONSFactory::getInstance()->createProducer(factory_info);
-        real_producer->start();
-    }
-
-    void HandleOKCallback()
-    {
-        Nan::HandleScope scope;
-
-        ons.real_producer = real_producer;
-        ons.initializing = false;
-        ons.inited = true;
-        ons.started = true;
-        callback->Call(0, 0);
-    }
-
-private:
-    ONSProducerV8& ons;
-    ONSFactoryProperty& factory_info;
-    Producer* real_producer;
-};
-
-class ProducerSendWorker : public Nan::AsyncWorker {
-public:
-    ProducerSendWorker(
-            Nan::Callback* callback,
-            ONSProducerV8& _ons,
-            string _topic,
-            string _tags,
-            string _key,
-            string _content,
-            int64_t _send_at) :
-        AsyncWorker(callback),
-        ons(_ons),
-        topic(_topic),
-        tags(_tags),
-        key(_key),
-        content(_content),
-        send_at(_send_at),
-
-        errored(false),
-        error_msg("")
-    {
-    }
-
-    ~ProducerSendWorker() {}
-
-    void Execute()
-    {
-        Message msg(topic.c_str(), tags.c_str(), content.c_str());
-        if(key != "")
-        {
-            msg.setKey(key.c_str());
-        }
-        
-        // delay...
-        if(send_at != -1)
-        {
-            msg.setStartDeliverTime(send_at);
-        }
-
-        uv_mutex_lock(&ons.mutex);
-        Producer* real_producer = ons.real_producer;
-
-        try
-        {
-            send_result = real_producer->send(msg);
-        }
-        catch(const ONSClientException& e)
-        {
-            error_msg = e.GetMsg();
-            errored = true;
-            uv_mutex_unlock(&ons.mutex);
-            return;
-        }
-
-        uv_mutex_unlock(&ons.mutex);
-    }
-
-    void HandleOKCallback()
-    {
-        Nan::HandleScope scope;
-
-        if(errored)
-        {
-            v8::Local<v8::Value> argv[1] = {
-                Nan::Error(error_msg.c_str())
-            };
-            callback->Call(1, argv);
-            return;
-        }
-
-        v8::Local<v8::Value> argv[2] = {
-            Nan::Undefined(),
-            Nan::New<v8::String>(send_result.getMessageId()).ToLocalChecked()
-        };
-        callback->Call(2, argv);
-    }
-
-private:
-    ONSProducerV8& ons;
-    string topic;
-    string tags;
-    string key;
-    string content;
-    
-    int64_t send_at;
-
-    SendResultONS send_result;
-    bool errored;
-    string error_msg;
-};
 
 ONSProducerV8::ONSProducerV8(string _producer_id, string _access_key, string _secret_key, ONSOptions _options) :
     producer_id(_producer_id),
@@ -266,7 +145,8 @@ NAN_METHOD(ONSProducerV8::Start)
 NAN_METHOD(ONSProducerV8::Stop)
 {
     ONSProducerV8* ons = ObjectWrap::Unwrap<ONSProducerV8>(info.Holder());
-    ons->Stop();
+    Nan::Callback* cb = new Nan::Callback(info[0].As<v8::Function>());
+    AsyncQueueWorker(new ProducerStopWorker(cb, *ons));
 }
 
 NAN_METHOD(ONSProducerV8::Send)
